+++ config/ac_oauth2.go
@@ -13,7 +13,7 @@ var (
 	_ OAuth2Authorization = &OAuth2AC{}
 )
 
-// OAuth2AC represents represents an oauth2 block for an OAuth2 client using the authorization code flow.
+// OAuth2AC represents an oauth2 block for an OAuth2 client using the authorization code flow.
 type OAuth2AC struct {
 	AccessControlSetter
 	AuthorizationEndpoint   string   `hcl:"authorization_endpoint"`
@@ -22,7 +22,6 @@ type OAuth2AC struct {
 	ClientSecret            string   `hcl:"client_secret"`
 	GrantType               string   `hcl:"grant_type"`
 	Name                    string   `hcl:"name,label"`
-	RedirectURI             string   `hcl:"redirect_uri"`
 	Remain                  hcl.Body `hcl:",remain"`
 	Scope                   *string  `hcl:"scope,optional"`
 	TokenEndpoint           string   `hcl:"token_endpoint"`
@@ -53,6 +52,7 @@ func (oa OAuth2AC) Schema(inline bool) *hcl.BodySchema {
 
 	type Inline struct {
 		Backend       *Backend `hcl:"backend,block"`
+		RedirectURI   string   `hcl:"redirect_uri"`
 		VerifierValue string   `hcl:"verifier_value"`
 	}
 
@@ -93,11 +93,7 @@ func (oa OAuth2AC) GetScope() string {
 	return *oa.Scope
 }
 
-func (oa OAuth2AC) GetRedirectURI() string {
-	return oa.RedirectURI
-}
-
-func (oa OAuth2AC) GetAuthorizationEndpoint() (string, error) {
+func (oa OAuth2AC) GetAuthorizationEndpoint(_ string) (string, error) {
 	return oa.AuthorizationEndpoint, nil
 }
 
@@ -110,6 +106,6 @@ func (oa OAuth2AC) GetTokenEndpointAuthMethod() *string {
 }
 
 // GetVerifierMethod retrieves the verifier method (ccm_s256 or state)
-func (oa OAuth2AC) GetVerifierMethod() (string, error) {
+func (oa OAuth2AC) GetVerifierMethod(_ string) (string, error) {
 	return oa.VerifierMethod, nil
 }