+++ config/oauth2.go
@@ -17,7 +17,7 @@ type OAuth2AS interface {
 // OAuth2AcAS represents the authorization server configuration for OAuth2 clients using the authorization code flow.
 type OAuth2AcAS interface {
 	OAuth2AS
-	GetAuthorizationEndpoint() (string, error)
+	GetAuthorizationEndpoint(uid string) (string, error)
 }
 
 // OidcAS represents the OIDC server configuration for OIDC clients.
@@ -41,18 +41,17 @@ type OAuth2Client interface {
 type OAuth2AcClient interface {
 	OAuth2Client
 	GetName() string
-	GetRedirectURI() string
 	// GetVerifierMethod retrieves the verifier method (ccm_s256, nonce or state)
-	GetVerifierMethod() (string, error)
+	GetVerifierMethod(uid string) (string, error)
 	GetBodyContent() *hcl.BodyContent
 }
 
 // OAuth2Authorization represents the configuration for the OAuth2 authorization URL function
 type OAuth2Authorization interface {
-	GetAuthorizationEndpoint() (string, error)
+	Inline
+	GetAuthorizationEndpoint(uid string) (string, error)
 	GetClientID() string
 	GetName() string
-	GetRedirectURI() string
 	GetScope() string
-	GetVerifierMethod() (string, error)
+	GetVerifierMethod(uid string) (string, error)
 }