+++ handler/file.go
@@ -8,12 +8,14 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/hashicorp/hcl/v2"
+
+	"github.com/avenga/couper/config/request"
 	"github.com/avenga/couper/config/runtime/server"
 	"github.com/avenga/couper/errors"
 	"github.com/avenga/couper/eval"
 	"github.com/avenga/couper/server/writer"
 	"github.com/avenga/couper/utils"
-	"github.com/hashicorp/hcl/v2"
 )
 
 const dirIndexFile = "index.html"
@@ -88,7 +90,7 @@ func (f *File) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 	}
 
 	if r, ok := rw.(*writer.Response); ok {
-		evalContext := req.Context().Value(eval.ContextType).(*eval.Context)
+		evalContext := req.Context().Value(request.ContextType).(*eval.Context)
 		r.AddModifier(evalContext, f.modifier)
 	}
 
@@ -103,7 +105,7 @@ func (f *File) serveDirectory(reqPath string, rw http.ResponseWriter, req *http.
 
 	if !strings.HasSuffix(reqPath, "/") {
 		if r, ok := rw.(*writer.Response); ok {
-			evalContext := req.Context().Value(eval.ContextType).(*eval.Context)
+			evalContext := req.Context().Value(request.ContextType).(*eval.Context)
 			r.AddModifier(evalContext, f.modifier)
 		}
 
@@ -122,7 +124,7 @@ func (f *File) serveDirectory(reqPath string, rw http.ResponseWriter, req *http.
 	defer file.Close()
 
 	if r, ok := rw.(*writer.Response); ok {
-		evalContext := req.Context().Value(eval.ContextType).(*eval.Context)
+		evalContext := req.Context().Value(request.ContextType).(*eval.Context)
 		r.AddModifier(evalContext, f.modifier)
 	}