+++ handler/spa.go
@@ -5,11 +5,13 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/hashicorp/hcl/v2"
+
+	"github.com/avenga/couper/config/request"
 	"github.com/avenga/couper/config/runtime/server"
 	"github.com/avenga/couper/errors"
 	"github.com/avenga/couper/eval"
 	"github.com/avenga/couper/server/writer"
-	"github.com/hashicorp/hcl/v2"
 )
 
 var (
@@ -60,7 +62,7 @@ func (s *Spa) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 	}
 
 	if r, ok := rw.(*writer.Response); ok {
-		evalContext := req.Context().Value(eval.ContextType).(*eval.Context)
+		evalContext := req.Context().Value(request.ContextType).(*eval.Context)
 		r.AddModifier(evalContext, s.modifier)
 	}